# nomlog-api/swagger.yaml
openapi: 3.0.3
info:
  title: Nomlog API
  description: API for meal logging and nutrition tracking
  version: 1.0.0
  contact:
    name: Nomlog API Support
servers:
  - url: http://localhost:3001/api/v1
    description: Local development server
  - url: https://nomlog-api.onrender.com/api/v1
    description: Production server
...
paths:
...
  /logs:
    get:
      tags:
        - Logs
      summary: Get meal logs for a date
      description: Retrieve meal logs for a specific date (defaults to today)
      operationId: getLogs
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Date in YYYY-MM-DD format (defaults to today)
          required: false
          schema:
            type: string
            format: date
            example: 2024-03-15
        - name: timezone
          in: query
          description: Timezone for date calculation (defaults to UTC)
          required: false
          schema:
            type: string
            example: America/New_York
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/MealLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Logs
      summary: Create meal log with nutrition analysis
      description: Create a new meal log by providing a natural language description. The API will analyze the meal and extract nutrition information.
      operationId: createLog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mealDescription
              properties:
                mealDescription:
                  type: string
                  description: Natural language description of the meal
                  example: I had a breakfast with 2 eggs, 2 slices of whole wheat toast with butter, and a cup of coffee with milk
                loggedAt:
                  type: string
                  format: date-time
                  description: When the meal was actually consumed (ISO 8601). If not provided, uses current time.
                  example: 2024-03-15T08:30:00Z
            examples:
              breakfast:
                summary: Breakfast example
                value:
                  mealDescription: I had a breakfast with 2 eggs, 2 slices of whole wheat toast with butter, and a cup of coffee with milk
              lunch:
                summary: Lunch example
                value:
                  mealDescription: Grilled chicken breast with brown rice and steamed broccoli
                  loggedAt: 2024-03-15T12:30:00Z
      responses:
        '201':
          description: Meal log created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Meal log created with nutrition analysis
                  data:
                    $ref: '#/components/schemas/MealLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /logs/{id}:
    get:
      tags:
        - Logs
      summary: Get meal log by ID
      description: Retrieve a specific meal log by its ID
      operationId: getLogById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Meal log ID
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Log retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  log:
                    $ref: '#/components/schemas/MealLog'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Logs
      summary: Update meal log timestamp
      description: Update when a meal was actually consumed
      operationId: updateLogTimestamp
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Meal log ID
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - loggedAt
              properties:
                loggedAt:
                  type: string
                  format: date-time
                  description: When the meal was actually consumed (ISO 8601)
                  example: 2024-03-15T08:30:00Z
      responses:
        '200':
          description: Log updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Meal log updated successfully
                  data:
                    $ref: '#/components/schemas/MealLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Unauthorized to update this meal log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Logs
      summary: Delete meal log
      description: Delete a meal log by ID
      operationId: deleteLog
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Meal log ID
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Log deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Meal log deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    MealLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          example: Breakfast
        description:
          type: string
          example: I had a breakfast with 2 eggs, 2 slices of whole wheat toast with butter, and a cup of coffee with milk
        total_nutrition:
          $ref: '#/components/schemas/Nutrition'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        logged_at:
          type: string
          format: date-time
          nullable: true
          description: When the meal was actually consumed
...
    Nutrition:
      type: object
      properties:
        calories:
          type: number
          example: 450
        fat:
          type: number
          example: 22
        protein:
          type: number
          example: 20
        carbohydrates:
          type: number
          example: 45
        fiber:
          type: number
          example: 5
        sugar:
          type: number
          example: 8
        sodium:
          type: number
          example: 400
        saturatedFat:
          type: number
          example: 10
        potassium:
          type: number
          example: 300
        cholesterol:
          type: number
          example: 200
        calcium:
          type: number
          example: 100
        iron:
          type: number
          example: 2
        vitaminA:
          type: number
          example: 500
        vitaminC:
          type: number
          example: 10
        vitaminD:
          type: number
          example: 50
        magnesium:
          type: number
          example: 50
    Ingredient:
      type: object
      properties:
        name:
          type: string
          example: Eggs
        servingAmount:
          type: number
          example: 2
        servingUnit:
          type: string
          example: eggs
        servingSizeGrams:
          type: number
          example: 100
        nutrition:
          $ref: '#/components/schemas/Nutrition'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (for 400 responses)
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'